MOSC-2716 / Implementar processo de LOGs para operações críticas.

+++ app/Http/Controllers/ParticipacaoSocialConferenciaController.php
@@ -2,12 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\Services\AuditService;
 use App\Services\Osc\ParticipacaoSocialConferenciaService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class ParticipacaoSocialConferenciaController extends Controller
 {
+    private $auditService;
     private $service;
 
     /**
@@ -18,6 +21,7 @@ class ParticipacaoSocialConferenciaController extends Controller
     public function __construct(ParticipacaoSocialConferenciaService $_service)
     {
         $this->service = $_service;
+        $this->auditService = new AuditService();
     }
 
     public function getAll()
@@ -73,8 +77,18 @@ class ParticipacaoSocialConferenciaController extends Controller
         try {
             $dados = $request->all();
 
+            $entidade = $this->service->store($dados);
+
+            if (!$entidade)
+            {
+                return response()->json(['Resposta' => 'Objeto não encontrado!'], Response::HTTP_OK);
+            }
+
+            $usuario = Auth::user();
+            $this->auditService->auditar('novoPartSocialConferencia', 'PartSocialConferencia', $entidade->id_conferencia, $usuario->id_usuario, 'criado', $entidade, $request->ip(), $entidade->id_osc);
+
             //Retorna novo registro
-            return response()->json($this->service->store($dados), Response::HTTP_OK);
+            return response()->json($entidade, Response::HTTP_OK);
         }
         catch (\Exception $e) {
             return $e->getMessage();
@@ -85,24 +99,40 @@ class ParticipacaoSocialConferenciaController extends Controller
         try {
             $dados = $request->all();
 
-            $ps_conferencia = $this->service->update($id, $dados);
+            $dados_old = $this->service->get($id);
 
-            if ($ps_conferencia)
+            $entidade = $this->service->update($id, $dados);
+
+            if (!$entidade)
             {
-                return response()->json(['Resposta' => 'Participação Social Conferência atualizado com sucesso!'], Response::HTTP_OK);
+                return response()->json(['Resposta' => 'Objeto não encontrado!'], Response::HTTP_OK);
             }
 
-            return $ps_conferencia;
+            $usuario = Auth::user();
+            $this->auditService->auditar('updatePartSocialConferencia', 'PartSocialConferencia', $id, $usuario->id_usuario, $dados_old, $dados, $request->ip(), $dados_old->id_osc);
+
+            return response()->json($dados, Response::HTTP_OK);
         }
         catch (\Exception $e) {
             return $e->getMessage();
         }
     }
 
-    public function delete($id_conferencia) {
+    public function delete($id_conferencia, Request $request) {
         try {
+
+            $dados_old = $this->service->get($id_conferencia);
+
+            if (!$dados_old)
+            {
+                return response()->json(['Resposta' => 'Objeto não encontrado!'], Response::HTTP_OK);
+            }
+
             if ($this->service->delete($id_conferencia))
             {
+                $usuario = Auth::user();
+                $this->auditService->auditar('deletePartSocialConferencia', 'PartSocialConferencia', $id_conferencia, $usuario->id_usuario, $dados_old, 'deletado', $request->ip(), $dados_old->id_osc);
+
                 return response()->json(['Resposta' => 'Participação Social Conferência deletado com sucesso!'], Response::HTTP_OK);
             }
         }
